backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.6.0
channels==4.1.0
requests==2.32.3
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
daphne==4.1.2
psycopg2-binary==2.9.10
twilio==9.4.6

black==24.8.0
isort==5.13.2
flake8==7.1.1
mypy==1.11.2

gunicorn==23.0.0
uvicorn==0.30.6